7-todo/backend/requirements.txt
asgiref==3.8.1
black==25.1.0
click==8.1.8
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
sqlparse==0.5.3
